refactor(graph): decompose graph.go into focused files [mache-godfile-graph]
Pure-move decomposition of internal/graph/graph.go (1420 → 159 LOC) across 8 files. No behavior changes.

New files:
- memstore.go (273) — struct + read methods
- memstore_write.go (376) — mutations + indexNode + ShiftOrigins
- memstore_refs.go (129) — refs/defs maps + accessors
- memstore_callees.go (190) — GetCallees + Go import parsing
- memstore_refsdb.go (207) — SQLite sidecar
- memstore_livegraph.go (44) — mtime tracking
- content_cache.go (80) — FIFO content cache extracted
- graph.go (159) — Graph interface + Node/ContentRef types

3 new behavioral tests in memstore_*_test.go. likeMatch deliberately left in sqlite_graph.go (composite + memstore call it cross-package-scope).

Coverage-gate (--rename-threshold 50): exit 0. 38 truly-new lines (memstore_livegraph.go + new file overhead), all covered.

Skeptic-agent verdict: APPROVE WITH COMMENT. Tests are real-behavior, file layout coherent, API preserved (all 38 MemoryStore signatures byte-equal), lock discipline clean, -race -count=5 clean, no test-only exports. One doc comment addressed in followup commit.

Closes mache-godfile-graph campaign work. Lands on feat/godfile-decomposition (PR #385's successor trunk).
